[FEATURE] Added LogBase class
* Added LogBase class to make it easier to use the logger
  in Python classes
