+++ pyproject.toml
@@ -4,7 +4,7 @@ build-backend = "setuptools.build_meta"
 
 [project]
 name = "uglylogger" # Required
-version = "0.7.1" # Required
+version = "0.8.0" # Required
 description = "An ugly, slow Logger class for python" # Optional
 readme = "README.md" # Optional
 requires-python = ">=3.10"
